Renames sponsors collection.
