+++ static/admin/config.yml
@@ -49,14 +49,14 @@ collections:
               {label: "ID", name: "id", widget: ncw-id, prefix: tier, hidden: true},
               {label: "Badge", name: "badge", widget: image, required: false},
               {label: "Lobby", name: "lobby", widget: object, fields: [
-                {label: "Button", name: button, widget: object, fields: [
-                  {label: "Text", name: "text", widget: string, required: false},
-                  {label: "Link", name: "link", widget: string, required: false}
-                ]},
                 {label: "Lobby Template", name: "lobby-template", widget: select, required: false, options: [
                   { label: "Big Images", value: "big-images" },
                   { label: "Small Images", value: "small-images" },
                   { label: "Horizontal Images", value: "horizontal-images" },
+                ]},
+                {label: "Button", name: button, widget: object, fields: [
+                  {label: "Text", name: "text", widget: string, required: false},
+                  {label: "Link", name: "link", widget: string, required: false}
                 ]},
               ]},
               {label: "Sponsor Page Template", name: "page-template", widget: select, options: [
@@ -202,16 +202,16 @@ collections:
       - {label: "Template Key", name: "templateKey", widget: "hidden", default: "custom-page"}
       - {label: Title, name: title, widget: string}
       - {label: "Body", name: body, widget: markdown}
-  - name: "sponsorData"
-    label: "Sponsor Data"
+  - name: "sponsorsData"
+    label: "Sponsors Data"
     editor:
       preview: false
     files:
-      - file: "src/content/sponsor.json"
-        label: "Sponsors"
-        name: "sponsor"
+      - file: "src/content/sponsors.json"
+        label: "Tier Sponsors"
+        name: "tierSponsors"
         fields:
-          - {label: "Sponsors", name: "tierSponsors", widget: list, summary: "{{tier[0].label}}", fields: [
+          - {label: "Tier Sponsors", name: "tierSponsors", widget: list, summary: "{{tier[0].label}}", fields: [
               {label: "Tier", name: "tier", widget: ncw-file-relation, collection: configurations, file: sponsorsTiers, target_field: tiers, id_field: id, display_fields: name},
               {label: "Sponsors", name: "sponsors", widget: list, fields: [
                 {label: "Name", name: "name", widget: string },
